microservices/db-manager/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.2.1
requests==2.32.3
gunicorn==23.0.0

